+++ Build.ps1
@@ -1,73 +1,16 @@
-param(
-    [String] $majorMinor = "1.8",  # 2.0
-    [String] $patch = "0",         # $env:APPVEYOR_BUILD_VERSION
-    [String] $customLogger = "",   # C:\Program Files\AppVeyor\BuildAgent\Appveyor.MSBuildLogger.dll
-    [Switch] $notouch
-)
+Push-Location $PSScriptRoot
 
-function Set-AssemblyVersions($informational, $assembly)
-{
-    (Get-Content assets/CommonAssemblyInfo.cs) |
-        ForEach-Object { $_ -replace """1.0.0.0""", """$assembly""" } |
-        ForEach-Object { $_ -replace """1.0.0""", """$informational""" } |
-        ForEach-Object { $_ -replace """1.1.1.1""", """$($informational).0""" } |
-        Set-Content assets/CommonAssemblyInfo.cs
-}
+if(Test-Path .\artifacts) { Remove-Item .\artifacts -Force -Recurse }
 
-function Install-NuGetPackages()
-{
-    nuget restore serilog-sinks-splunk.sln
-}
+& dotnet restore
 
-function Invoke-MSBuild($solution, $customLogger)
-{
-    if ($customLogger)
-    {
-        msbuild "$solution" /verbosity:minimal /p:Configuration=Release /logger:"$customLogger"
-    }
-    else
-    {
-        msbuild "$solution" /verbosity:minimal /p:Configuration=Release
-    }
-}
+$revision = @{ $true = $env:APPVEYOR_BUILD_NUMBER; $false = 1 }[$env:APPVEYOR_BUILD_NUMBER -ne $NULL];
 
-function Invoke-NuGetPackProj($csproj)
-{
-    nuget pack -Prop Configuration=Release -Symbols $csproj
-}
+Push-Location src/Serilog.Sinks.Splunk
 
-function Invoke-NuGetPackSpec($nuspec, $version)
-{
-    nuget pack $nuspec -Version $version -OutputDirectory ..\..\
-}
+& dotnet pack -c Release -o ..\..\.\artifacts --version-suffix=$revision
+if($LASTEXITCODE -ne 0) { exit 1 }    
 
-function Invoke-NuGetPack($version)
-{ 
-    pushd .\src\Serilog.Sinks.Splunk
-    Invoke-NuGetPackSpec "Serilog.Sinks.Splunk.nuspec" $version
-    popd
-}
 
-function Invoke-Build($majorMinor, $patch, $customLogger, $notouch)
-{
-    $package="$majorMinor.$patch"
-
-    Write-Output "Building Serilog.Sinks.Splunk $package"
-
-    if (-not $notouch)
-    {
-        $assembly = "$majorMinor.0.0"
-
-        Write-Output "Assembly version will be set to $assembly"
-        Set-AssemblyVersions $package $assembly
-    }
-
-    Install-NuGetPackages
-    
-    Invoke-MSBuild "serilog-sinks-splunk.sln" $customLogger
-
-    Invoke-NuGetPack $package
-}
-
-$ErrorActionPreference = "Stop"
-Invoke-Build $majorMinor $patch $customLogger $notouch
+Pop-Location
+Pop-Location