+++ sample/Serilog.Sinks.Splunk.Sample/IConfigure.cs
@@ -1,7 +1,0 @@
-namespace Serilog.Sinks.Splunk.Sample
-{
-    internal interface IConfigure
-    {
-        void Configure();
-    }
-}