+++ src/Serilog.Sinks.Splunk/Sinks/Splunk/EventCollectorRequest.cs
@@ -1,4 +1,20 @@
+// Copyright 2016 Serilog Contributors
+// 
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// 
+//     http://www.apache.org/licenses/LICENSE-2.0
+// 
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+
 using System;
+using System.Globalization;
 using System.Net.Http;
 using System.Text;
 
@@ -34,7 +50,7 @@ namespace Serilog.Sinks.Splunk
 
             if (time > 0)
             {
-                jsonPayLoad = jsonPayLoad + @",""time"":" + time;
+                jsonPayLoad = jsonPayLoad + @",""time"":" +  time.ToString(CultureInfo.InvariantCulture);
             }
 
             jsonPayLoad = jsonPayLoad + "}";
@@ -49,11 +65,17 @@ namespace Serilog.Sinks.Splunk
 
     internal class EventCollectorRequest : HttpRequestMessage
     {
-        internal EventCollectorRequest(string splunkHost, string jsonPayLoad)
+        internal EventCollectorRequest(string splunkHost, string jsonPayLoad, string uri ="services/collector")
         {
+            var hostUrl = $@"{splunkHost}/{uri}";
 
+            if(splunkHost.Contains("services/collector"))
+            {
+                hostUrl = $@"{splunkHost}";
+            }
+            
             var stringContent = new StringContent(jsonPayLoad, Encoding.UTF8, "application/json");
-            RequestUri = new Uri(splunkHost);
+            RequestUri = new Uri(hostUrl);
             Content = stringContent;
             Method = HttpMethod.Post;
         }