+++ src/Serilog.Sinks.Splunk/Sinks/Splunk/UdpSink.cs
@@ -1,4 +1,7 @@
-﻿// Copyright 2014 Serilog Contributors
+﻿#if UDP
+
+
+// Copyright 2016 Serilog Contributors
 // 
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -87,7 +90,7 @@ namespace Serilog.Sinks.Splunk
         /// <param name="renderTemplate">If true, the message template will be rendered</param>
         [Obsolete("Use the overload accepting a connection info object instead. This overload will be removed.", false)]
         public UdpSink(IPAddress hostAddress, int port, IFormatProvider formatProvider = null, bool renderTemplate = true)
-            : this(new SplunkUdpSinkConnectionInfo(hostAddress, port),formatProvider, renderTemplate)
+            : this(new SplunkUdpSinkConnectionInfo(hostAddress, port), formatProvider, renderTemplate)
         {
         }
 
@@ -115,4 +118,43 @@ namespace Serilog.Sinks.Splunk
             _socket = null;
         }
     }
+
+    /// <summary>
+    /// Defines connection info used to connect against Splunk
+    /// using UDP.
+    /// </summary>
+    public class SplunkUdpSinkConnectionInfo
+    {
+        /// <summary>
+        /// Splunk host.
+        /// </summary>
+        public IPAddress Host { get; }
+
+        /// <summary>
+        /// Splunk port.
+        /// </summary>
+        public int Port { get; }
+
+        /// <summary>
+        /// Creates an instance of <see cref="SplunkUdpSinkConnectionInfo"/> used
+        /// for defining connection info for connecting using UDP against Splunk.
+        /// </summary>
+        /// <param name="host">Splunk host.</param>
+        /// <param name="port">Splunk UDP port.</param>
+        public SplunkUdpSinkConnectionInfo(string host, int port) : this(IPAddress.Parse(host), port) { }
+
+        /// <summary>
+        /// Creates an instance of <see cref="SplunkUdpSinkConnectionInfo"/> used
+        /// for defining connection info for connecting using UDP against Splunk.
+        /// </summary>
+        /// <param name="host">Splunk host.</param>
+        /// <param name="port">Splunk UDP port.</param>
+        public SplunkUdpSinkConnectionInfo(IPAddress host, int port)
+        {
+            Host = host;
+            Port = port;
+        }
+    }
 }
+
+#endif